+++ android/samples/sample-multi-view/src/main/java/com/google/android/filament/multiview/MainActivity.kt
@@ -65,6 +65,7 @@ class MainActivity : Activity() {
     private lateinit var view1: View
     private lateinit var view2: View
     private lateinit var view3: View
+    private lateinit var view4: View
     // We need skybox to set the background color
     private lateinit var skybox: Skybox
     // Should be pretty obvious :)
@@ -119,11 +120,15 @@ class MainActivity : Activity() {
         view1 = engine.createView()
         view2 = engine.createView()
         view3 = engine.createView()
+        view4 = engine.createView()
 
         view0.setName("view0");
         view1.setName("view1");
         view2.setName("view2");
         view3.setName("view3");
+        view4.setName("view4");
+
+        view4.blendMode = View.BlendMode.TRANSLUCENT;
 
         skybox =  Skybox.Builder().build(engine);
         scene.skybox = skybox
@@ -136,11 +141,13 @@ class MainActivity : Activity() {
         view1.camera = camera
         view2.camera = camera
         view3.camera = camera
+        view4.camera = camera
 
         view0.scene = scene
         view1.scene = scene
         view2.scene = scene
         view3.scene = scene
+        view4.scene = scene
     }
 
     private fun setupScene() {
@@ -373,6 +380,7 @@ class MainActivity : Activity() {
         engine.destroyView(view1)
         engine.destroyView(view2)
         engine.destroyView(view3)
+        engine.destroyView(view4)
         engine.destroySkybox(skybox)
         engine.destroyScene(scene)
         engine.destroyCameraComponent(camera.entity)
@@ -411,6 +419,9 @@ class MainActivity : Activity() {
                     skybox.setColor(0.0f, 0.0f, 1.0f, 1.0f);
                     renderer.render(view3)
 
+                    skybox.setColor(0.0f, 0.0f, 0.0f, 0.0f);
+                    renderer.render(view4)
+
                     renderer.endFrame()
                 }
             }
@@ -443,6 +454,7 @@ class MainActivity : Activity() {
             view1.viewport = Viewport(width / 2, 0,          width / 2, height / 2)
             view2.viewport = Viewport(0,         height / 2, width / 2, height / 2)
             view3.viewport = Viewport(width / 2, height / 2, width / 2, height / 2)
+            view4.viewport = Viewport(width / 4, height / 4, width / 2, height / 2)
         }
     }